Fix timezone issue with weekly recurrences
When a date/time string (Such as `2014-03-02`, `14:10:00`, or `020314`)
is passed to `strtotime`, `strtotime` assumes that the string is
formatted to local convention and timezone, unless indicated otherwise.

For example, when given `14:00:00` on a server with a default timezone
of `Asia/Tokyo`, strtotime will assume that that is 14:00 local time in
Tokyo.

On the line modified we provide a similar string, but as it comes from a
unix timestamp passed through `gmdate`, the resultant string handed to
`strtotime` is utc and not local time. By adding a `z` to the end of the
string, we tell `strtotime` to treat it as utc.

(An alternative solution would be to change the `gmdate` to `date`, but
that would have php translating utc->local for `date`, then local->utc
for `strtotime`)