Fix 3 "Failures" reported by `composer test`
Code was generating a date-time string that was essentially UTC, but formatting
it in such a way that was causing `iCalDateToUnixTimestamp` to interpret it as
a "floating" local time.

Solution is not to format a integer into a string and then interpret it into a
different integer.

A better solution might be to maintain a php DateTime object instead of an integer
timestamp: bumping up the days, months, weeks or years as necessary, so constantly
converting between integers, strings, and the odd DateTime object is unnecessary.

However, rewritting much of the event recurrence handling code is beyond the scope
of the pull request this commit is destined for.