(Deindent 1/2) Invert the logic of two if-statements.
The first now provides an early return if there's no work to do,
the second an early `continue` if the condition is not made.

As there's no `else` or anything after either `if` statements,
there's no point in having a 708+ line indented codeblock.

(Indenting change is split across two commits to be more easy to comprehend at a later date.)