Remove method `trimToRecurrenceCount` as no longer necessary
Before the commit prior to this one, we added the quantity of EXDATEs
onto `$countLimit` (formerly `$countOrig`). This was causing the
`ProcessRecurrences` method to create more occurrences than necessary.

Now we no longer add the EXDATE quantity, the method should no longer
overproduce, thus we shouldn't need to cull occurrences.