Permit arbitrary lines to be fed into Recurrence Tests
Currently, the test framework in `RecurrenceTests.php` permit three
ical lines to be passed when creating a `VEVENT` to test. (Expected to
be `DTSTART`, `DTEND`, and `RRULE`)

Some tests only use two lines (typically skipping `DTEND`); whilst it
would also be nice to be pass more lines, so as to build tests with
arbitrary quantities of `EXDATE`s, `RDATE`s, etc.

The `rfc5545RecurrenceExamplesTest.php` file already uses this same
methodology to permit more than three lines to be added to a test
`VEVENT`.