Upgrade `phpunit/phpunit` dependency to 5.x.x
The version used previously was using the core php function `each()` which was
deprecated in `php 7.2` and removed for `php 8.0`. As such, those running
webservers with `php 8` were no longer able to run tests.

We now rely on a slightly more recent version of `phpunit` which also bumps our
minimum supported version of `php` to `5.6`.