Handle case of RRULE:COUNT=1 (any FREQ)
Outlook, as reported in #293, may in some cases reduce `COUNT` to `1`.

Whilst not disallowed in the RFC5545 spec (or any of its errata to date),
having a `COUNT` of `1` has no practical purpose that I can see.

We currently only check we've reached the count-limit *after* we've generated
the first few recurrences of an event and - more pressingly - *after* we've
appended the first candidate to the array of generated recurrences.

In this case `PHP_MAX_INT` could be replaced with some suitable default number,
akin to the current `defaultSpan` property used to put a limit on UNTIL.
But that's something for another PR.